fix: grant tests
